use hash id based on request path since module id may be inconsistent between client and server builds.
