+++ lib/addStylesClient.js
@@ -14,6 +14,8 @@ if (typeof DEBUG !== 'undefined' && DEBUG) {
 	) }
 }
 
+var listToStyles = require('./listToStyles')
+
 /*
 type StyleObject = {
 	id: number;
@@ -45,10 +47,10 @@ var noop = function () {}
 // tags it will allow on a page
 var isOldIE = typeof navigator !== 'undefined' && /msie [6-9]\b/.test(navigator.userAgent.toLowerCase())
 
-module.exports = function (list, _isProduction) {
+module.exports = function (parentId, list, _isProduction) {
 	isProduction = _isProduction
 
-  var styles = listToStyles(list)
+  var styles = listToStyles(parentId, list)
   addStylesToDom(styles)
 
   return function update (newList) {
@@ -60,7 +62,7 @@ module.exports = function (list, _isProduction) {
       mayRemove.push(domStyle)
     }
     if (newList) {
-      var newStyles = listToStyles(newList)
+      var newStyles = listToStyles(parentId, newList)
       addStylesToDom(newStyles)
     }
     for (var i = 0; i < mayRemove.length; i++) {
@@ -85,19 +87,19 @@ function addStylesToDom (styles /* Array<StyleObject> */) {
         domStyle.parts[j](item.parts[j])
       }
       for (; j < item.parts.length; j++) {
-        domStyle.parts.push(addStyle(item.parts[j], item.id, j))
+        domStyle.parts.push(addStyle(item.parts[j]))
       }
     } else {
       var parts = []
       for (var j = 0; j < item.parts.length; j++) {
-        parts.push(addStyle(item.parts[j], item.id, j))
+        parts.push(addStyle(item.parts[j]))
       }
       stylesInDom[item.id] = { id: item.id, refs: 1, parts: parts }
     }
   }
 }
 
-function listToStyles (list) {
+function listToStyles (parentId, list) {
   var styles = []
   var newStyles = {}
   for (var i = 0; i < list.length; i++) {
@@ -108,8 +110,10 @@ function listToStyles (list) {
     var sourceMap = item[3]
     var part = { css: css, media: media, sourceMap: sourceMap }
     if (!newStyles[id]) {
+      part.id = parentId + ':0'
       styles.push(newStyles[id] = { id: id, parts: [part] })
-    } else			{
+    } else {
+      part.id = parentId + ':' + newStyles[id].parts.length
       newStyles[id].parts.push(part)
     }
   }
@@ -123,9 +127,9 @@ function createStyleElement () {
   return styleElement
 }
 
-function addStyle (obj /* StyleObjectPart */, id, index) {
+function addStyle (obj /* StyleObjectPart */) {
   var update, remove
-  var styleElement = document.querySelector('style[data-vue-ssr-id~="' + id + ':' + index + '"]')
+  var styleElement = document.querySelector('style[data-vue-ssr-id~="' + obj.id + '"]')
   var hasSSR = styleElement != null
 
   // if in production mode and style is already provided by SSR,
