+++ lib/addStylesServer.js
@@ -1,4 +1,6 @@
-module.exports = function (list, isProduction) {
+var listToStyles = require('./listToStyles')
+
+module.exports = function (parentId, list, isProduction) {
   if (typeof __VUE_SSR_CONTEXT__ !== 'undefined') {
     var context = __VUE_SSR_CONTEXT__
     var styles = context._styles
@@ -14,34 +16,13 @@ module.exports = function (list, isProduction) {
       })
     }
 
-    list = listToStyles(list)
+    list = listToStyles(parentId, list)
     if (isProduction) {
       addStyleProd(styles, list)
     } else {
       addStyleDev(styles, list)
     }
   }
-}
-
-function listToStyles (list) {
-  var styles = []
-  var newStyles = {}
-  for (var i = 0; i < list.length; i++) {
-    var item = list[i]
-    var id = item[0]
-    var css = item[1]
-    var media = item[2]
-    var sourceMap = item[3]
-    var part = { css: css, media: media, sourceMap: sourceMap }
-    if (!newStyles[id]) {
-      part.id = id + ':0'
-      styles.push(newStyles[id] = { id: id, parts: [part] })
-    } else {
-      part.id = id + ':' + newStyles[id].parts.length
-      newStyles[id].parts.push(part)
-    }
-  }
-  return styles
 }
 
 // In production, render as few style tags as possible.
@@ -52,12 +33,13 @@ function addStyleProd (styles, list) {
     for (var j = 0; j < parts.length; j++) {
       var part = parts[j]
       // group style tags by media types.
-      var style = styles[part.media]
+      var id = part.media || 'default'
+      var style = styles[id]
       if (style) {
         style.ids.push(part.id)
         style.css += '\n' + part.css
       } else {
-        styles[part.media || 'default'] = {
+        styles[id] = {
           ids: [part.id],
           css: part.css,
           media: part.media
