+++ package.json
@@ -9,6 +9,7 @@
   },
   "license": "MIT",
   "dependencies": {
+    "hash-sum": "^1.0.2",
     "loader-utils": "^0.2.7"
   }
 }
